Fixed python3 only bug and rectified failing stix-validation tests
Fixed a python 3 only bug in the verify_content_is_valid functtion in
the opentaxii/taxii/utils.py file. Fixed failing tests in the test_inbox.py
file too. These test failures deserve special mention. In the Cosive
modifications we aligned the TAXII status messages better with the TAXII
standard, so that a failed validation returns a failure status to the client.
This ensures that the client knows that there was a problem with that content.
We also slightly modified the way processing was done, so that in this version,
if there is a single invalid content_block, then ALL content_blocks are ignored,
the idea being that the TAXII client should fix up the TAXII message messaage
and should resend it again, fixed.

NOTE! THe above is different from the way that the official OpenTAXII responds.
The official OpenTAXII looks at all contentn_block, and as long as they are
ok, it will ingest the valid content_blocks, silently ignoring the invalid content.
There is no indication given to the client that something went wrong.

This is what we changed in our version. The client is notified that something is
wrong, and no content_blocks from that TAXII message are ingested. The problem
with our approach is that users may be expecting the content to still be ingested.
This subtle change of behaviour may impact some other downstream customers, so we
may revisit this change shortly.